+++ akka-http-testkit/src/main/scala/akka/http/javadsl/testkit/RouteTest.scala
@@ -10,7 +10,7 @@ import scala.concurrent.duration._
 import akka.stream.ActorMaterializer
 import akka.http.scaladsl.server
 import akka.http.javadsl.model.HttpRequest
-import akka.http.javadsl.server.{ Route, Directives }
+import akka.http.javadsl.server.{ AllDirectives, Route, Directives }
 import akka.http.impl.util.JavaMapping.Implicits._
 import akka.http.impl.server.RouteImplementation
 import akka.http.scaladsl.model.HttpResponse
@@ -19,7 +19,7 @@ import akka.actor.ActorSystem
 import akka.event.NoLogging
 import akka.http.impl.util._
 
-abstract class RouteTest {
+abstract class RouteTest extends AllDirectives {
   implicit def system: ActorSystem
   implicit def materializer: ActorMaterializer
   implicit def executionContext: ExecutionContext = system.dispatcher
