+++ akka-http-tests-java8/src/test/java/akka/http/javadsl/server/HandlerBindingTest.java
@@ -7,6 +7,7 @@ package akka.http.javadsl.server;
 import akka.http.scaladsl.model.HttpRequest;
 import org.junit.Test;
 import akka.http.javadsl.testkit.*;
+import akka.http.javadsl.server.values.*;
 import static akka.http.javadsl.server.Directives.*;
 
 public class HandlerBindingTest extends JUnitRouteTest {
