+++ akka-http/src/main/scala/akka/http/impl/server/RouteImplementation.scala
@@ -4,6 +4,8 @@
 
 package akka.http.impl.server
 
+import akka.http.javadsl.server.values.{ PathMatcher, BasicUserCredentials }
+
 import scala.language.implicitConversions
 import scala.annotation.tailrec
 import scala.collection.immutable
